// src/api/generated/models/consolidationStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ConsolidationStatus = typeof ConsolidationStatus[keyof typeof ConsolidationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ConsolidationStatus = {
  DRAFT: 'DRAFT',
  PLANNED: 'PLANNED',
  IN_TRANSIT: 'IN_TRANSIT',
  CUSTOMS: 'CUSTOMS',
  DELIVERED: 'DELIVERED',
} as const;